Updated playerPrenguin.prefab + Updated obstacleScene.unity + Updated Scripts
Updated playerPrenguin.prefab
Updated the playerPrenguin.prefab, simplified the collissions to use only one Capsule Collider.

Updated obstacleScene.unity
Updated the obstacleScene.unity file. Added more test obstacles to see how the player reacts.

Updated Scripts
Updated the NightGuardAI.cs Script to include the visionMesh variable (Used to detect the player, add the Cone Mesh from the Flashlight to use it)

Updated OrbitCameraController.cs to check if the player is sliding (If they are, it stops the player from moving the camera)

Updated the PlayerControllerRigid to fix the sliding mechanic (Adjusted the time from 5 seconds to 2 seconds, updated the maxSpeed to 8.0f when sliding, stopped the player from being able to sprint when sliding)