Map Prefabs Set Up
All Sheleves and Map Parts (Pillars and Doorframes) now have their own prefab. Existing assets on the map will be replaced by the prefab tomorrow or when Darcy adds in his version of the map. But for now, everything has a clean prefab so people can click and drag them in.