Updated Prefabs + Started Working on Improved Camera Clipping
Updated Prefabs
Updated some Objects to use the ObjectMovementScript instead of an Animator for Movement, so they come out much smoother and are more controlable. Also, the pickups have had their colliders changed to Spheres so they don't push the player around.

Started Working on Improved Camera Clipping
I have started to work on a Camera Clipping system to get the points from the corners of the Near Clip Plane and use those for the new distance rather then just a point from the center of the camera.